requirements.txt
beautifulsoup4==4.5.3
bleach==1.4.3
ConcurrentLogHandler==0.9.1
contextlib2==0.5.4
dj-database-url==0.4.2
Django==1.10.7
django-appconf==1.0.2
django-compressor==2.1.1
django-debug-toolbar==1.7
django-modelcluster==3.0.1
django-redis==4.7.0
django-taggit==0.22.0
django-treebeard==4.1.0
djangorestframework==3.6.2
elasticsearch==5.3.0
Embedly==0.5.0
html5lib==0.9999999
httplib2==0.10.3
Markdown==2.6.2
olefile==0.44
Pillow==4.1.0
pkg-resources==0.0.0
psycopg2==2.7.1
Pygments==2.2.0
pytz==2017.2
raven==6.0.0
rcssmin==1.0.6
redis==2.10.5
requests==2.13.0
rjsmin==1.0.12
six==1.10.0
sqlparse==0.2.3
Unidecode==0.4.20
urllib3==1.20
wagtail==1.9
wagtail-markdown==0.4.1
wagtailmodeladmin==2.5.7
webencodings==0.5.1
whitenoise==3.3.0
Willow==0.4
